#comment ANTON refactoring and debugging adaptive
